Move casting to cast
