+++ src/api/cast/barrel.ts
@@ -1,4 +1,5 @@
 export * from './Airplay';
 export * from './CastAPI';
+export * from './CastConfiguration';
 export * from './CastState';
 export * from './Chromecast';
